Fix the elevation of IntrinsicSizeFrameLayout when the layout is MATCH_PARENT
The IntrinsicSizeFrameLayout is shown in the child screen that is not a full screen. Even though the IntrinsicSizeFrameLayout is set MATCH_PARENT to be a fullscreen. The edge line would be shown on IntrinsicSizeFrameLayout.

Fix the issue that the elevation of IntrinsicSizeFrameLayout is not reset to show the edge line  when the layout params are set MATCH_PARENT and it isn't the real fullscreen.


Bug: 332639317
Test: manual test
Change-Id: Iee49ffd94ac72ef063a8478735232829d06469d8