[SetupDesign] Fix the down button position incorrect and buttons stick together issue.
The root case should be a timing issue, so the solution we make seconday button as GONE at the init monent, then relay on the state to show up secondary button.

For down button shows up case, the secondary is not neceressary to show up. Ones the content cannot scrollable, then we show the secondary button.

Bug: 424728773
Test: manual
Change-Id: Ifdb15af31ff0df736e367e9be9e57a18a0afa354